+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- Change default value for `butler.graphql.namespace`
 - Refactor butler-health config.
 - Refactor butler-audit config.
 - Merge package and application "butler" config.