+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- Refactor butler-audit config.
 - Merge package and application "butler" config.
 
 ## [0.13.0] - 2021-09-28