+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.14.0] - 2021-09-30
+
 ### Changed
 - Change default value for `butler.graphql.namespace`
 - Refactor butler-health config.